--- src/client/component/classComboBox.tsx
@@ -1,17 +1,0 @@
-import * as React from 'react';
-import TextField from '@mui/material/TextField';
-import Autocomplete from '@mui/material/Autocomplete';
-
-import allCourses from '../data/allCourses';
-
-export default function ClassComboBox() {
-  return (
-    <Autocomplete
-      disablePortal
-      options={allCourses}
-      sx={{ width: 300 }}
-      renderInput={(params) => <TextField {...params} label="Classes" />}
-    />
-  );
-}
-
