+++ src/client/css/About.css
@@ -1,0 +1,41 @@
+.main {
+    position: relative;
+    width: 905px;
+    height: 680px;
+    margin: 0 auto;  /* Center the main container horizontally */
+    background: white;
+    border: 8px solid #1e1e1e;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 18px;
+    color: black;
+    display: flex;
+    flex-direction: column;  /* Stack the contents vertically */
+    justify-content: flex-start;  /* Ensure content starts at the top */
+    align-items: center;  /* Center content horizontally */
+}
+
+.about-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    padding: 20px;
+    text-align: center;
+    margin-top: 60px; 
+}
+
+.about-container h1 {
+    margin-bottom: 10px;
+}
+
+.about-container p {
+    max-width: 600px;
+    line-height: 1.5;
+}
+
+.back-button {
+    position: absolute !important;
+    top: 20px;
+    left: 20px;
+    z-index: 999;  /* Ensure it's on top of other elements */
+}
