+++ src/client/css/FileUploadButton.css
@@ -1,0 +1,19 @@
+.editor-button {
+    background-color: #554481;
+    color: white;
+}
+
+.hidden-file-input {
+    display: none;
+}
+
+.sr-only {
+    position: absolute;
+    width: 1px;
+    height: 1px;
+    padding: 0;
+    margin: -1px;
+    overflow: hidden;
+    clip: rect(0, 0, 0, 0);
+    border: 0;
+}
