+++ src/client/main.tsx
@@ -1,14 +1,14 @@
-import "./index.css";
+import "./css/index.css";
 import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
 
 import React from "react";
 import ReactDOM from "react-dom/client";
 
-import App from "./App";
-import Home from "./Home";
-import Login from "./login";
-import About from "./About";
-import Editor from "./Editor";
+// import App from "./routes/App";
+import Home from "./routes/Home";
+import Login from "./routes/Login";
+import About from "./routes/About";
+import Editor from "./routes/Editor";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
@@ -16,7 +16,7 @@ ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
       <Routes>
         <Route path="/" element={<Login />} />
         <Route path="/Home" element={<Home />} />
-        <Route path="/App" element={<App />} />
+        {/*<Route path="/App" element={<App />} />*/}
         <Route path="/Login" element={<Login />} />
         <Route path="/About" element={<About />} />
         <Route path="/Editor" element={<Editor />} />
